feat(schematics): implement Siemens migration rules and refactor style mappings
