+++ projects/schematics/siemens-migration/ts-import-migration-rule.ts
@@ -1,0 +1,101 @@
+/**
+ * Copyright (c) Siemens 2016 - 2025
+ * SPDX-License-Identifier: MIT
+ */
+import { Tree, SchematicContext, Rule, chain } from '@angular-devkit/schematics';
+import * as ts from 'typescript';
+
+import { getImportNodes, getSymbols, discoverSourceFiles } from '../utils';
+import { findComponentImportPath } from './mappings';
+import { MigrationOptions, Migrations } from './model';
+
+export const importMigrationRule = (_options: MigrationOptions): Rule => {
+  return (tree: Tree, context: SchematicContext) => {
+    const rules: Rule[] = [];
+
+    const sourceFiles = discoverSourceFiles(tree, context, _options);
+
+    for (const filePath of sourceFiles) {
+      const migrations = collectMigrationImports(filePath, tree, context);
+      const { imports, toRemoveImports } = migrations;
+
+      if (imports.size === 0 && toRemoveImports.length === 0) {
+        continue;
+      }
+
+      const rule = rewriteImportsInFile(filePath, migrations);
+      rules.push(rule);
+    }
+
+    return chain([...rules]);
+  };
+};
+
+const collectMigrationImports = (
+  filePath: string,
+  tree: Tree,
+  _context: SchematicContext
+): Migrations => {
+  const content = tree.read(filePath);
+  const imports = new Map<string, string[]>();
+
+  if (!content) {
+    return { imports, toRemoveImports: [] };
+  }
+
+  const simplImports = getImportNodes(filePath, content.toString(), '@simpl/');
+  const toRemoveImports: ts.ImportDeclaration[] = [];
+
+  for (const node of simplImports) {
+    const symbols = getSymbols(node);
+    const importPath = node.moduleSpecifier.getText().replace(/['"]/g, '');
+    let toRemove = false;
+    for (const symbol of symbols) {
+      const symbolName = symbol.name.getText();
+      const newImportPath = findComponentImportPath(symbolName, importPath);
+      if (!newImportPath) {
+        continue;
+      }
+
+      toRemove = true;
+      const migration = imports.get(newImportPath);
+      if (!migration) {
+        imports.set(newImportPath, [symbolName]);
+      } else {
+        migration.push(symbolName);
+      }
+    }
+
+    if (toRemove) {
+      toRemoveImports.push(node);
+    }
+  }
+
+  return { imports, toRemoveImports };
+};
+
+const rewriteImportsInFile = (filePath: string, migrations: Migrations): Rule => {
+  return (tree: Tree): Tree => {
+    const recorder = tree.beginUpdate(filePath);
+    const start = migrations.toRemoveImports.at(0)?.getFullStart() ?? 0;
+
+    // Remove old imports
+    for (const importDecl of migrations.toRemoveImports) {
+      recorder.remove(importDecl.getFullStart(), importDecl.getFullWidth());
+    }
+
+    // Add new imports
+    const importPaths = Array.from(migrations.imports.keys()).sort((a, b) => a.localeCompare(b));
+
+    for (const path of importPaths) {
+      const symbols = migrations.imports.get(path)!;
+      recorder.insertLeft(
+        start,
+        `\nimport { ${symbols.sort((a, b) => a.localeCompare(b)).join(', ')} } from '${path}';`
+      );
+    }
+
+    tree.commitUpdate(recorder);
+    return tree;
+  };
+};
